+++ lang/da/navig.php
@@ -2,12 +2,10 @@
 
 return [
 
-    'login' => 'Log ind',
-    'register' => 'Ny konto',
-    'register.email' => 'Ny konto med e-mail',
+    'login' => 'Log ind eller Opret konto',
+    'pin-login' => 'Log ind eller Opret konto',
     'users.index' => 'Brugere',
     'users.show' => 'Brugeroplysninger',
-    'pin-login' => 'Login med e-mail',
-    'pin-login.verify' => 'Kodebekræftelse',
+    'pin-login.verify' => 'PIN-bekræftelse',
 
 ];